+++ src/router/registry/Task.vue
@@ -24,19 +24,36 @@
     </h3>
     <br />
     <router-view/>
+    <json-tree
+      v-if="registryTask"
+      :data="registryTask"
+      rootObjectKey="task"
+      maxDepth="3"></json-tree>
   </div>
 </template>
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
 import Entity from '@/components/Entity'
+import JsonTree from '@/components/JsonTree'
 
 export default {
   components: {
-    Entity
+    Entity,
+    JsonTree
   },
   computed: {
-    ...mapGetters(['task', 'worker'])
+    ...mapGetters(['task', 'worker', 'registryWorkers']),
+    registryTask: function () {
+      let task = null
+      const worker = this.registryWorkers
+        .find(el => el.name === this.worker.name)
+      if (worker) {
+        task = worker.tasks.find(el2 => el2.name === this.task.name)
+      }
+
+      return task
+    }
   },
   methods: {
     ...mapActions(['query'])