+++ test/unit/setup.js
@@ -1,6 +1,10 @@
 import Vue from 'vue'
+import VueRouter from 'vue-router'
 import BootstrapVue from 'bootstrap-vue'
+import TreeView from 'vue-json-tree-view'
 
 Vue.config.productionTip = false
 
+//Vue.use(VueRouter)
 Vue.use(BootstrapVue)
+//Vue.use(TreeView)